chore: Move new iterators into old iterator place.
