+++ src/map.rs
@@ -4,8 +4,9 @@ use louds_rs::Louds;
 mod longest_prefix_iter;
 mod trie;
 mod trie_builder;
-
-pub mod clone;
+pub mod postfix_iter;
+pub mod search_iter;
+pub mod prefix_iter;
 
 /// A trie for sequences of the type `Label`; each sequence has an associated `Value`.
 pub struct Trie<Label, Value> {
