--- src/map/clone/mod.rs
@@ -1,7 +1,0 @@
-//! A trie map stores a value with each word or key.
-pub mod postfix_iter;
-pub mod search_iter;
-pub mod prefix_iter;
-// mod trie;
-
-// pub use trie::{Trie,TrieBuilder};
